win: Next phase of GN toolchain
This imports some goop from gyp/gn/chromium to locate the
system-installed Visual Studio (rather than requiring it in the path).

This is necessary to be able to do both x86 and x64 builds in the same
shell.

The toolchain is not yet templated to support that, and always builds as
x64. However, this does now avoid the need to run vcvarsall before
building.

BUG=crashpad:79

Change-Id: I9669b0f65ee154fe35769453844a1d48cf84b74c