Enable thread-safe statics when building with GCC and clang
This is effective for macOS and Linux (including Android).

This is the same as upstream
https://codereview.chromium.org/2667723002/. /Zc:threadSafeInit- was
never added to mini_chromium as part of bug chromium:587210, so there’s
no need to remove it. MSVC supports thread-safe static initialization
since VS2015, GCC has supported it since version 4.0 (2005), and clang
has probably supported it since the beginning.

BUG=chromium:686866

Change-Id: I06cc737a9f27891c1bceda7f1251ace9f1d7cb29