android: Support “unified headers” with GCC build
Previously, the __ANDROID_API__ macro was provided by
<android/api-level.h>. With unified headers (expected to become the
default in the next NDK release and the sole option in the subsequent
release), this macro is not set properly by headers. When buliding with
Clang, the standalone toolchain’s clang and clang++ wrappers set the
macro properly. GCC isn’t accounted for in this way, so the build system
must assume the responsibility of setting it.

This change allows __ANDROID_API__ to be defined based on the
android_sdk_version GYP variable, taken from the GYP variable of the
same name in Chromium when it used GYP. For an example of how this might
be used, see the dependent
https://chromium-review.googlesource.com/458021/.

https://android.googlesource.com/platform/ndk/+/master/docs/UnifiedHeaders.md
https://android.googlesource.com/platform/ndk/+/ndk-r14/CHANGELOG.md

Bug: crashpad:30
Change-Id: Ia910fe11c5b5a5b872249ff7389b24a210373967