android: Recognize Bionic’s semantics for strerror_r()
In previous versions of POSIX, the claim goes, the specification of
strerror_r() was ambiguous, and it was unclear whether it returned an
error number on failure, or returned nonzero while setting errno to an
error number. The standard, if it ever was actually ambiguous on this
point, no longer is. The former behavior is required, but that doesn’t
change the fact that noncompliant libraries like Bionic continue to
implement the latter.

https://android.googlesource.com/platform/bionic/+/26bc9c64d5b0cad921e3070a4f94fa04e1077d90/libc/bionic/strerror_r.cpp#52

Adapt to Bionic’s quirk: recognize both failure mechanisms by treating
positive return values as error numbers, and negative return values as
failure with the error number in errno. Valid error numbers are always
positive, and Bionic returns -1 on failure.

Chromium has an unnecessarily complex way of achieving the same
behavior.

This does not affect the use of strerror_r() with glibc, which exposes
yet another distinct and nonstandard interface.

Bug: crashpad:30
Change-Id: I5c0fe5d7631d224a02d3d3d19457d953422945a8