linux: Switch between x86 and x86_64 with the target_arch GYP variable
When targeting Linux on x86, with proper toolchain support, it’s
possible to switch the target in a native (non-cross) build by using
-m32 and -m64. Use the GYP target_arch variable to direct this, as was
done for Mac in 6c49a5924dea.

This is only done for OS=="linux", and notably not OS=="android", which
must always be configured as a cross build.

Bug: crashpad:30
Change-Id: I4a0b0d0ffd9911a18c155fa9a991ccd31ce08d25