fuchsia: Look for the toolchain and SDK in per-build-host directories
I like to share a single Crashpad checkout between my non-virtual
machine and some virtual machines. Downloaded toolchains, which vary by
build host configuration, must go in paths named for the build host.
(Chromium doesn’t do this, and it bugs me.)

Rather than looking for the Fuchsia toolchain in
third_party/fuchsia/clang, this looks for it in
third_party/fuchsia/clang/{mac,linux}-amd64.

The Fuchsia SDK is only published on cipd for linux-amd64, but the
sysroot that it contains ought to be perfectly functional on any
suitably-equipped build host, so this also uses that package
unconditionally.

Bug: crashpad:196
Change-Id: Ibd0660eb7ef82a2e6b7e6adf113edb80fabe6036