gn, mac: Honor mac_sdk_min, sysroot, and mac_deployment_target
This is bit more flexible (and probably straightforward) than how
Chromium deals with things. The SDK is selected based on the first
non-empty GN argument out of sysroot, mac_sdk, and mac_sdk_min.
Regardless of how the SDK is selected, if mac_sdk is non-empty, the SDK
must be that exact version, and if mac_sdk_min is non-empty, the SDK
must be at least that version. It’s possible for all three to be empty,
which will result in the system’s default SDK being chosen, without any
additional constraints. This is the current default.

Similarly, mac_deployment_target may be empty, which allows the
toolchain to choose the deployment target. The current default is to
explicitly specify 10.9 as the deployment target.

Unlike Chromium, this ought to work on systems with only the Command
Line Tools package installed, but no Xcode. Also unlike Chromium, no
provision has been made (yet?) for the hermetic tools package.

The sysroot logic for macOS and Fuchsia has been consolidated, although
there’s probably more room for improvement.

Bug: crashpad:79
Change-Id: I1df56921bd68934b6c4671707ca47842b58047fe