gn: Don’t use .rsp files; rationalize tool descriptions
.rsp files were omitted from the Windows build (where they’re probably
more likely to be useful) for simplicity, since mini_chromium and
Crashpad shouldn’t need them. They were left in place for link steps on
other platforms, though. This removes the use of .rsp files, which
should make it easier to debug the build.

Some of the tool descriptions were a little wacky. Some were just plain
wrong, and others were just a little wrong. Rather than trying to match
Chromium step-for-step, this standardizes on a set of meaningful and
recognizable names not likely to be confused for anything else: CC, CXX,
OBJC, OBJCXX, ASM, AR, SOLINK_MODULE, LINK, STAMP, and COPY.

The Objective-C steps are defined on macOS only.

Bug: crashpad:79
Change-Id: I1160279f06ccbb663087be494795588f4a637843