Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Joshua, what do you think about this?
> 
> I’m not so sure about not using the system’s default compiler in the default configuration.

Seems reasonable, I'm easy either way.

I could also switch it around to default to just whatever "clang++" it finds, and then have the bots pass in an argument to override.

The obvious drawback is that the bots don't build what a default get-and-build does, so that has the potential to cause a bit of confusion.

Patch-set: 3
Reviewer: Gerrit User 1144260 <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>