Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> Is this the right approach? mini_chromium intentionally maintains its own idea of “is_posix” called “mini_chromium_is_posix” so that we get to make the call on things like this.
> 
> (Initially, this was created because Fuchsia’s build system didn’t even have anything called is_posix, so we needed our own thing.)
> 
> +scottmg for Fuchsia-related things, although he’s digging out.

I should have put a bit more details in the CL description.
My reasoning was that we will also want to change is_posix to false for mini_chromium for consistency with the chromium. But now that I think some more about it, if we want to do that, maybe we can just do a single switch here after the crashpad CL has landed?
This way, we can consider Fuchsia as a non-POSIX platform in crashpad if it is compiled as part of chromium or as a stand-alone.

Patch-set: 3
Reviewer: Gerrit User 1116524 <1116524@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>