Update patch set 3
Patch Set 3:

Right now, we consider the Chromium Fuchsia build sets |is_posix| to true. The long-term plan is to switch that to false. Regarding the implications:
-OS_POSIX will no longer be defined. Right now, we define a lot of things like
  #if defined(OS_POSIX) && !defined(OS_FUCHSIA)
  These will become redundant when we switch. I will clean these up post-switch.
-*_posix files will keep being included by default for the time being. Technically, this works as a global "exclusion filter". Each target defines a set of source files, including *_win, *_android, *_posix, etc. And a global exclusion filter removes the *_posix files if the build is not posix. Since this is too large a change to do at the same time as switching |is_posix| to false, we will keep the *_posix files by default in Fuchsia for now. After the switch, I will go back and remove all *_posix files from the default source set, add them explicitly for Posix and/or Fuchsia and then remove the global exclusion filter, which should also speed up non-posix builds.

I wrote this doc about switching:
https://docs.google.com/document/d/13Ft6KsZ12C97QUdNn7lwCjLmX2WVJfnhkhyoTXiC9JI/edit?usp=sharing

For the CL at hand, mini_chromium is not compiled as part of the Chromium build but is consumed by Crashpad, which will have different values for |crashpad_is_posix| depending on whether Crashpad is built as part of Chromium or as a stand-alone. It is not a big problem, we can handle both cases easily but it makes the logic in the Crashpad build files more complex than it needs to be. This is why I believe we should switch mini_chromium to set |is_posix| to false for Fuchsia too.

Hope that helps.

Patch-set: 3
Reviewer: Gerrit User 1116524 <1116524@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>