Various toolchain churn for Linux and Fuchsia
The clang that the Fuchsia team ships now requires -ldl and -lpthread
for host (Linux) binaries, but that must currently be put manually on
the command line.

Additionally, while we were force linking fdio on Fuchsia, we weren't
including zircon. But fdio won't work without zircon, so include it too.

Bug: crashpad:30, crashpad:196
Change-Id: I13fba13acd0380bb18143524336160096bac10a6