Update patch set 5
Patch Set 5:

(3 comments)

Gerrit is showing my comments to me out of order, but rest assured that I mean for you to read the comment on line 72 before the other two.

This may be the best solution in that it gives a stronger guarantee of order and in that it’s most similar to what Chromium does.

Patch-set: 5
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>