[ios/mac] Use clonefile instead of hardlink to populate bundles
Use `cp -c` to copy file into the bundle which uses clonefile instead
of hardlink. This is as efficient in speed and disk usage. This also
ensures that files in the bundle are distinct from any other files
(which is important when running EG2 tests via Xcode as it tries to
overwrite some files with `ditto` which fails if source and target
are hardlinked together).

Taken from crrev.com/c/2133930

Change-Id: Ic3874ce3654a6923452e9be7ea8c9af2bbd9ba49