mini_chromium: Fix a gn check issue.
public_deps should be used for deps which appear in a public .h,
since anyone including that .h can't help but take the transitive
dependency.

This is really not that important and I only sent this CL because
I thought I was fixing a different problem (which I'm not).
Still, here it is, if you want it.

Change-Id: Icc46a082351426b10ad18fe83abc91f1305ba9f9