Fix Fuchsia failing builds
Fuchsia relies on this repository. It broke after we updated CBU8_NEXT
because there is an implicit downcast from an int to an uint8_t in
utf_string_conversion_utils.cc.


The latest Chromium code for ReadUnicodeCharacter has been updated
and no longer has this compilation issue (I verified this in godbolt).  I
double checked and there seem to be no other references to CBU8_NEXT in this
repository so this should be the only issue introduced that we need to fix.

Bug: chromium:1484644
Change-Id: Iab311dce9ef8e84b1df6a0d6a9c698cac2943fb7